+++ servant-util-beam-pg/CHANGES.md
@@ -3,3 +3,4 @@ Unreleased
 
 * Add implementation of sorting.
 * Add implementation of pagination.
+* Add implementation of filtering.